+++ src/main/java/roomescape/time/repository/ReservationTimeDao.java
@@ -1,5 +1,6 @@
 package roomescape.time.repository;
 
+import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
 
@@ -58,5 +59,13 @@ public class ReservationTimeDao {
 
     public void delete(long id) {
         jdbcTemplate.update("DELETE FROM RESERVATION_TIME WHERE ID = ?", id);
+    }
+
+    public List<ReservationTime> available(LocalDate parse, long themeId) {
+        String query="SELECT rt.id, rt.start_at " +
+                "FROM reservation_time rt " +
+                "LEFT JOIN reservation r ON rt.id = r.time_id AND r.date = ? AND r.theme_id = ? " +
+                "WHERE r.id IS NULL";
+        return jdbcTemplate.query(query,reservationTimeRowMapper,parse,themeId);
     }
 }
